restructure utils functions
